src/main/webapp/app/entities/cobranca-empresa/delete/cobranca-empresa-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { CobrancaEmpresaService } from '../service/cobranca-empresa.service';

import { CobrancaEmpresaDeleteDialogComponent } from './cobranca-empresa-delete-dialog.component';

describe('CobrancaEmpresa Management Delete Component', () => {
  let comp: CobrancaEmpresaDeleteDialogComponent;
  let fixture: ComponentFixture<CobrancaEmpresaDeleteDialogComponent>;
  let service: CobrancaEmpresaService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [CobrancaEmpresaDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(CobrancaEmpresaDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(CobrancaEmpresaDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(CobrancaEmpresaService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
